Separates the installer into 2 smaller ones
