+++ src/api/v1beta1/properties.go
@@ -21,7 +21,7 @@ import (
 	"net/url"
 	"strings"
 
-	"github.com/Dynatrace/dynatrace-operator/src/dtclient"
+	"github.com/Dynatrace/dynatrace-operator/src/dtclient/types"
 	"github.com/pkg/errors"
 	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
@@ -280,21 +280,21 @@ func (dk *DynaKube) Tokens() string {
 	return dk.Name
 }
 
-func (dk *DynaKube) CommunicationHostForClient() dtclient.CommunicationHost {
-	return dtclient.CommunicationHost(dk.Status.CommunicationHostForClient)
+func (dk *DynaKube) CommunicationHostForClient() types.CommunicationHost {
+	return types.CommunicationHost(dk.Status.CommunicationHostForClient)
 }
 
-func (dk *DynaKube) ConnectionInfo() dtclient.ConnectionInfo {
-	return dtclient.ConnectionInfo{
+func (dk *DynaKube) ConnectionInfo() types.ConnectionInfo {
+	return types.ConnectionInfo{
 		CommunicationHosts: dk.CommunicationHosts(),
 		TenantUUID:         dk.Status.ConnectionInfo.TenantUUID,
 	}
 }
 
-func (dk *DynaKube) CommunicationHosts() []dtclient.CommunicationHost {
-	var communicationHosts []dtclient.CommunicationHost
+func (dk *DynaKube) CommunicationHosts() []types.CommunicationHost {
+	var communicationHosts []types.CommunicationHost
 	for _, communicationHost := range dk.Status.ConnectionInfo.CommunicationHosts {
-		communicationHosts = append(communicationHosts, dtclient.CommunicationHost(communicationHost))
+		communicationHosts = append(communicationHosts, types.CommunicationHost(communicationHost))
 	}
 	return communicationHosts
 }
