+++ src/installer/common/consts.go
@@ -1,0 +1,5 @@
+package common
+
+const (
+	AgentConfDirPath = "agent/conf"
+)
