--- src/installer/config.go
@@ -1,9 +1,0 @@
-package installer
-
-import (
-	"github.com/Dynatrace/dynatrace-operator/src/logger"
-)
-
-var (
-	log = logger.NewDTLogger().WithName("oneagent-installer")
-)
