+++ src/installer/image/config.go
@@ -1,0 +1,24 @@
+package image
+
+import (
+	"path/filepath"
+
+	dtcsi "github.com/Dynatrace/dynatrace-operator/src/controllers/csi"
+	"github.com/Dynatrace/dynatrace-operator/src/logger"
+)
+
+const (
+	rawPolicy = `{
+		"default": [
+			{
+				"type": "insecureAcceptAnything"
+			}
+		]
+	}
+	`
+)
+
+var (
+	CacheDir = filepath.Join(dtcsi.DataPath, "cache")
+	log      = logger.NewDTLogger().WithName("oneagent-image-installer")
+)
