+++ src/installer/image/destination_test.go
@@ -1,4 +1,4 @@
-package installer
+package image
 
 import (
 	"testing"
