+++ src/installer/symlink/symlink_test.go
@@ -1,4 +1,4 @@
-package installer
+package symlink
 
 import (
 	"path/filepath"
