+++ src/installer/url/config.go
@@ -1,0 +1,13 @@
+package url
+
+import (
+	"github.com/Dynatrace/dynatrace-operator/src/logger"
+)
+
+var (
+	log = logger.NewDTLogger().WithName("oneagent-url-installer")
+)
+
+const (
+	VersionLatest = "latest"
+)
