+++ src/installer/zip/gzip.go
@@ -1,4 +1,4 @@
-package installer
+package zip
 
 import (
 	"archive/tar"
@@ -9,76 +9,10 @@ import (
 	"strings"
 
 	"github.com/klauspost/compress/gzip"
-	"github.com/klauspost/compress/zip"
 	"github.com/spf13/afero"
 )
 
-func extractZip(fs afero.Fs, sourceFile afero.File, targetDir string) error {
-	if sourceFile == nil {
-		return fmt.Errorf("file is nil")
-	}
-
-	fileInfo, err := sourceFile.Stat()
-	if err != nil {
-		return fmt.Errorf("unable to determine file info: %w", err)
-	}
-
-	reader, err := zip.NewReader(sourceFile, fileInfo.Size())
-	if err != nil {
-		return fmt.Errorf("failed to open ZIP file: %w", err)
-	}
-
-	_ = fs.MkdirAll(targetDir, 0755)
-
-	for _, file := range reader.File {
-		if err := extractFileFromZip(fs, targetDir, file); err != nil {
-			return err
-		}
-	}
-
-	return nil
-}
-
-func extractFileFromZip(fs afero.Fs, targetDir string, file *zip.File) error {
-	path := filepath.Join(targetDir, file.Name)
-
-	// Check for ZipSlip: https://snyk.io/research/zip-slip-vulnerability
-	if !strings.HasPrefix(path, filepath.Clean(targetDir)+string(os.PathSeparator)) {
-		return fmt.Errorf("illegal file path: %s", path)
-	}
-
-	mode := file.Mode()
-
-	// Mark all files inside ./agent/conf as group-writable
-	if file.Name != agentConfPath && isAgentConfFile(file.Name) {
-		mode |= 020
-	}
-
-	if file.FileInfo().IsDir() {
-		return fs.MkdirAll(path, mode)
-	}
-
-	if err := fs.MkdirAll(filepath.Dir(path), mode); err != nil {
-		return err
-	}
-
-	dstFile, err := fs.OpenFile(path, os.O_WRONLY|os.O_CREATE|os.O_TRUNC, mode)
-	if err != nil {
-		return err
-	}
-	defer func() { _ = dstFile.Close() }()
-
-	srcFile, err := file.Open()
-	if err != nil {
-		return err
-	}
-	defer func() { _ = srcFile.Close() }()
-
-	_, err = io.Copy(dstFile, srcFile)
-	return err
-}
-
-func extractGzip(fs afero.Fs, sourceFilePath, targetDir string) error {
+func ExtractGzip(fs afero.Fs, sourceFilePath, targetDir string) error {
 	targetDir = filepath.Clean(targetDir)
 	log.Info("extracting tar gzip", "source", sourceFilePath, "destinationDir", targetDir)
 
@@ -158,12 +92,4 @@ func extractGzip(fs afero.Fs, sourceFilePath, targetDir string) error {
 			log.Info("skipping special file", "name", header.Name)
 		}
 	}
-}
-
-// Checks if a file is under /agent/conf
-// Different archive file implementations return different output, this function handles the differences
-// zip.File.Name == "path/to/file"
-// tar.Header.Name == "./path/to/file"
-func isAgentConfFile(fileName string) bool {
-	return strings.HasPrefix(fileName, "./"+agentConfPath) || strings.HasPrefix(fileName, agentConfPath)
 }
