+++ src/processmoduleconfig/config.go
@@ -1,0 +1,9 @@
+package processmoduleconfig
+
+import (
+	"github.com/Dynatrace/dynatrace-operator/src/logger"
+)
+
+var (
+	log = logger.NewDTLogger().WithName("processmoduleconfig")
+)
