+++ src/standalone/run_test.go
@@ -6,15 +6,16 @@ import (
 	"testing"
 
 	"github.com/Dynatrace/dynatrace-operator/src/dtclient"
+	"github.com/Dynatrace/dynatrace-operator/src/dtclient/types"
 	"github.com/Dynatrace/dynatrace-operator/src/installer"
 	"github.com/spf13/afero"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
 
-var testProcessModuleConfig = dtclient.ProcessModuleConfig{
+var testProcessModuleConfig = types.ProcessModuleConfig{
 	Revision: 0,
-	Properties: []dtclient.ProcessModuleProperty{
+	Properties: []types.ProcessModuleProperty{
 		{
 			Section: "test",
 			Key:     "test",
@@ -129,7 +130,7 @@ func TestInstallOneAgent(t *testing.T) {
 		runner := createMockedRunner(t)
 		runner.dtclient.(*dtclient.MockDynatraceClient).
 			On("GetProcessModuleConfig", uint(0)).
-			Return(&dtclient.ProcessModuleConfig{}, fmt.Errorf("BOOM"))
+			Return(&types.ProcessModuleConfig{}, fmt.Errorf("BOOM"))
 		runner.installer.(*installer.InstallerMock).
 			On("UpdateProcessModuleConfig", BinDirMount, &testProcessModuleConfig).
 			Return(nil)
